Event Messaging Changes
events no longer require a name be passed to dispatcher as the type of
event now comes from the message itself.

AbacuraMessage is now a dataclass, as are all subclasses of it.

Events have more standardized naming, with core.* for abacura-core
events and lok.* for LOK ones:
  msdp_value_CHARACTER_NAME is now core.msdp.CHARACTER_NAME

I think I got all the dispatchers and message types.