Move from session.telnet_client to TelnetPlugin
This means moving the connection point into the badly named
session.launch_screen(), so that we can connect after Plugins are
loaded.  Also tweak to autobuf to prevent it from firing immediately on
login, but I'll revisit that again, we likely need an "in-game" state.

MSDP is loaded by the session, and passed as a handler list to the
connection.  This will make loading additional telnet sub protocol bits
easier in the future (GMCP, MSSP, etc).